+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## dev ()
+
+* Handle multiple `:test_paths` in `mix.exs`
+
 ## v0.3.0 (2022-04-07)
 
 * Don't report `defstruct` as not covered